refactor: move CI-policy tests out of cmd and split the protocol commands
The release and release-automation workflow contracts lived in the `cmd`
package, so the CLI package's tests depended on `.github/workflows`.
They move to a test-only `internal/releasepolicy`, alongside the
existing `internal/architecture`, and the architecture allowlist now
names both so neither can quietly grow product code.

The tag-validation test asserted the step's regex literals and that the
step does not call `grep`. That froze one implementation of the rule
without checking the rule, so it moves to executing the step: a matrix
of tags that must reach the job holding the publishing secrets and tags
that must not, including leading zeros, build metadata, a shell
substitution, and a refs/ path. Every structural and supply-chain
assertion is kept as it was — no secrets in validate or test, no manual
environment approval on release, no fallback to a token that suppresses
tag workflows, no comparison against a moving latest release, and
observations that stay advisory. The injection test gains the same
Windows skip as its siblings, since it also needs bash.

`cmd/run.go` defined `protocol submit|answer|skip|resume` while
`cmd/protocol.go` held only the group and `material`. The four
subcommands and the three next-variant helpers only they use now live
with the group.

The architecture guide lists the three test-only packages, which its
package table omitted.
